Add audio/flake to the Port's Collection.
The Flake encoder project was started as a way for the author to learn more
about lossless audio encoding. The author chose the FLAC format because it is
open and well-documented. Although the FLAC reference encoder is also free and
open-source, the author's goal was to implement the standard from scratch as an
independent project. Flake is licensed under the GNU LGPL.

Once Flake progressed to a usable state, the author submitted the encoder to be
included as part of FFmpeg. It was included and improved upon by other FFmpeg
developers. Since then, Ithe author has added more features to Flake and
improved the encoding speed. The author plan to add some of these features into
FFmpeg as the encoder progresses.

There are both benefits and drawbacks to using Flake as a FLAC encoder
alternative. In its current state, Flake gives, on average, slightly better
compression than the FLAC reference encoder and is about twice as fast.
However, it does not currently implement all the advanced features such as
tagging, cuesheets, 24-bit audio, embedded pictures, and ogg encapsulation.
The author hopes to add these to Flake in future releases.

WWW: http://flake-enc.sourceforge.net

Approved by:	bdrewery (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@312940 35697150-7ecd-e111-bb59-0022644237b5