- Update to 3.16
This fixes a regression with non-pkgng usage that would invoke
the bootstrapper when upgrading ports. My testing did not
cover systems with the bootstrapper that did not use pkgng, so
this was missed.

The impact of this was just a hung build during deinstall.

Reported by:	mandree
Pointyhat to:	bdrewery


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@315585 35697150-7ecd-e111-bb59-0022644237b5