Update to 6.4.4
* new option in file config.tcl indicating whether GPSMan data can be
appended to an existing file when saving; to be used with care as a
file with conflicting time offsets can result that will not be
loaded by GPSMan; asked by Zvi Grauer.
* explicit support for the Garmin Street Pilot I, with thanks to
Martin Espinoza who sent the protocols list.

In addition, I tried to clean up the port a wee bit:
* I removed a line in the Makefile that had been commented-out
Wed Jul 26 08:32:31 2006 UTC (r168793).
* I changed the USE_TK_RUN specification to permit the use of
Tcl/Tk 8.5 (and tested that it brings in 8.5 if that's
available, and that gpsman works with it).

PR:		ports/178112
Submitted by:	dhw (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@316463 35697150-7ecd-e111-bb59-0022644237b5