www/py-praw: Python Reddit API Wrapper [NEW PORT]
PRAW, an acronym for "Python Reddit API Wrapper", is a python package
that allows for simple access to reddit's API. PRAW aims to be as easy
to use as possible and is designed to follow all of reddit's API rules.

You have to give a useragent that follows the rules, everything else is
handled by PRAW so you needn't worry about violating them.

WWW: http://praw.readthedocs.org


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@322084 35697150-7ecd-e111-bb59-0022644237b5