Update to 5.3.27
Please Note: This will be the last regular release of the PHP 5.3 series.
All users of PHP are encouraged to upgrade to PHP 5.4 or PHP 5.5. The PHP
5.3 series will receive only security fixes for the next year.

I will maintain PHP 5.3 until the first not fixed vulnerability is found.
At that point I will deprecate the port with a 2 months notice, if no
other volunteer is found in that time the port will be removed.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@322936 35697150-7ecd-e111-bb59-0022644237b5