Add new port: lang/gnat_util
There are a number of GNAT tools that have to be build using the same
exact sources as the compiler that it serves.  Two examples of this are
GPRbuild and ASIS.  The latter is an important component of several other
language semantic tools.

Accordingly, a separate port is being created to build a static library
that need these compiler sources.  Incidentally, diverging sources is the
reason devel/gprbuild-aux stopped building after lang/gcc-aux was upgraded.
With this new scheme, this type of failure should no longer occur.

This is a dependency of the upcoming lang/asis port, and must be specified
as a build dependency by any port that needs ASIS.

Approved by: 	bapt (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@323388 35697150-7ecd-e111-bb59-0022644237b5