devel/adasdl: Move to Ada framework
Currently this port has a hard dependency on lang/gnat.  This has been
replaced with USES+=ada and minor changes to incorporate it into the Ada
framework with a PORTREVISION bump.

The makefile header was trimmed, spaces were replaced with tabs, the
leading article was removed from the COMMENT, USES_GMAKE was converted
to USES and GMAKE was redefined in order to access the environment
variables.

This is a very old program, and the data types defined in it in are
too small for amd64.  Redports confirms it will not build on this platform
so ONLY_FOR_ARCHS=i386 is set as well.

Approved by:	bapt & culot (mentors)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@323754 35697150-7ecd-e111-bb59-0022644237b5