Introduce a new option BOOTSTRAP that allows switching between a full bootstrap of GCC and just running a simple build.
Bootstrapping is the default upstream, and for very stable and old
flavors of GCC we disable this by default since that is a huge win
in terms of build time of the port.  No change in default behavior
for this port -- for the time being we continue to bootstrap.

Update to the 20130822 snapshot of GCC 4.8.2.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@325321 35697150-7ecd-e111-bb59-0022644237b5