sysutils/ciso: Mark broken, deprecated and expired
The port is broken because the distfile can't be fetched due to upstream
site disappearing.  However, development has ceased for 6 years and it is
obsolete since similar projects such as cisoplus are still being developed.
Hopefully marking it deprecated and due for removal will provoke cisoplus
to be made into a port.

PR:		181558
Approved by:	rodrigo (maintainer)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@325790 35697150-7ecd-e111-bb59-0022644237b5