Add a new bsd.sanity.mk.
This file is responsible to do some sanity checking on port, as well as doing some sanity checking on users configuration
New macro for developer: WARNING (this will issue a warning to the normal users can be used to warn the user about deprecating feature that will be unsupported soon.
New macro to put in your make.conf for Maintainers: DEVELOPER=yes this macro will activate warning on usage of deprecated features and errors on now unspported usage of macros.

Reviewed by:	bdrewery


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@327576 35697150-7ecd-e111-bb59-0022644237b5