Mark x11-toolkits/py-kivy as broken under i386 without -DWITH_NEW_XORG.
For some unknown reason builds fail with:

  Undefined symbol "glBlendEquationSeparate"

When compiling under i386 for the old libGL.  This error does not make
sense given that the symbol is present in libGL and that compilation works
under amd64.  The only solution found is to update to a newer libGL however
there does not appear to be any change to the symbol signature to suggest
why such an update works.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@329608 35697150-7ecd-e111-bb59-0022644237b5