textproc/pdftohtml: Resolve conflict with poppler-utils
The conflict between this unmaintained port and poppler-utils can be
problematic.  Ports should be using poppler-utils rather than this one
as a build dependency.  This port's package list is 1-line long, so
lets just rename the program so that it can coexist with the version
from poppler-utils.  The name changes from "pdftohtml" to "pdf-to-html"
along with a PORTREVISION bump.

Minor cleanup to pet portlint was alson done.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@329645 35697150-7ecd-e111-bb59-0022644237b5