Add devel/py-singledispatch:
This library brings functools.singledispatch from Python 3.4 to
Python 2.6-3.3.

PEP 443 proposed to expose a mechanism in the functools standard
library module in Python 3.4 that provides a simple form of generic
programming known as single-dispatch generic functions.

This library is a backport of this functionality to Python 2.6 - 3.3.

WWW: https://pypi.python.org/pypi/singledispatch


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@335219 35697150-7ecd-e111-bb59-0022644237b5