security/suricata: Remove unsupported DAG (Endace Capture) OPTION
This OPTION is non-functional as it requires hardware support and libdag from
Endace, which is not available in, nor recommended to be built via the ports
tree.

This OPTION also incorrectly added CONFIGURE_ARGS without adding any
LIB_DEPENDS, which broke configure: during build when the option was enabled.

Reported by:	mat (via pkg-fallout, via IRC)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@335661 35697150-7ecd-e111-bb59-0022644237b5