Unlike your common unilateral falling block games, 'Double Cross' implements a bidirectional paradigm expanding the genre in both dimension and difficulty.
Blocks fall from the top and fly in from the side settling in a joined
"play area". Deletions occur when rows of ten blocks are completed.
A horizontal row of 10 will cause the blocks to fall down, and a vertical row
of 10 will cause the blocks to "fall" to the right.
If a vertical row is completed during a vertical drop or a horizontal row is
completed during a horizontal drop the corresponding deletion will not occur
until the next turn. This can and will lead to non-intuitive results.
Focus on the vertical alone and you will die from horizontal negligence and
vice versa.

WWW: https://code.google.com/p/double-cross/

PR:		ports/171603
Submitted by:	nemysis (self)
Approved by:	pawel (mentor)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@336720 35697150-7ecd-e111-bb59-0022644237b5