games/exult: Update to latest 1.5.x snapshot (Unbreaks F10+)
The previous version of Exult, 1.2, is considered ancient by the
developers.  Morever, it's c++ was horrid, requiring around 3 dozen
patches just to build on equally ancient gcc 4.5.

Exult has not been released yet, but the snapshots are considered
superior even to the last RC release, 1.4.9rc.  The C++ code builds
without patches on both gcc 4.7 (DragonFly) and clang from F10 (verified
by Redports)

The Exult Studio was added as an option, it's on by default.  I did not
add the Mod or usecode compiler.  If somebody wants those, they can be
added as options as well.  While updating, support stage.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@339906 35697150-7ecd-e111-bb59-0022644237b5