Properly set license when using the demosaic packs.
According to the port's README.demosaic-packs, when the demosaic packs are
used the software must be licensed under either the GPLv2 or the GPLv3,
depending on the packs being used (if both packs are used, the GPLv3 should
take precedence).

This patch changes the way LICENSE is set when those options are set: the
new licenses are exclusive with the others.

PR:		ports/185119
Submitted by:	rakuco@
Approved by:	maintainer timeout (29 days)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@340649 35697150-7ecd-e111-bb59-0022644237b5