lang/gnatdroid-armv(5,7): Use lang/gcc-aux as base, not lang/gnat-aux
After a few additional patches for gcc-aux, gnatdroid-armv5 (and -armv7)
build and execute well when built and based on gcc 4.7-based gcc-aux.
The previous version was based on gcc 4.6-based gnat-aux.  This was the
final port requiring gnat-aux, and now its part of the Ada framework.

Running acats using an ASUS Transformer XF101 (Android 4.0.3) as a target
results in a nearly perfect result.  Only test cb1010a fails, which is
a failure to detect running out of stack space (-fstack-check).  I intend
to investigate further, but as the only problem, gnatdroid remains a
highly capable compiler despite the small regression.  Potentially
zero-cost exceptions were gained as a trade-off.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@340860 35697150-7ecd-e111-bb59-0022644237b5