lang/gnat-aux: Mark deprecated in favor of lang/gcc-aux
Now that the gnatdroid-arm ports have been rebased on lang/gcc-aux, there
is no port that requires lang/gnat-aux.  Since it can't build most of the
Ada framework, there's no anticipated reason to use it over the newer,
gcc4.7-based gcc-aux.  Leave it around just in case, but let folks fooled
by "gnat" in name that it's not the latest Ada compiler.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@341146 35697150-7ecd-e111-bb59-0022644237b5