Add GMGPL to license database
The GNAT Modified General Public Licence is an older license that is
commonly used with Ada programs, specifically to address legal quirks
with the Ada generics feature.

It is meant to be used as "multi" licence in combination with GPL or
GPLv2.  It is not typically seen combined with GPLv3 to my knowledge.

Approved by:	portmgr (bapt)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344409 35697150-7ecd-e111-bb59-0022644237b5