- Add p5-Test-Bits 0.02
Test::Bits provides a single subroutine, bits_is(), for testing binary data.

This module is quite similar to Test::BinaryData and Test::HexString in concept.
The difference is that this module shows failure diagnostics in a different way,
and has a slightly different calling style. Depending on the nature of the data
you're working with, this module may be easier to work with.

WWW: http://search.cpan.org/dist/Test-Bits/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344482 35697150-7ecd-e111-bb59-0022644237b5