Turn devel/kyua into a meta-port and add devel/kyua-cli.
The existing devel/kyua package should have been named devel/kyua-cli to
match the upstream structure.  Do the rename now and make devel/kyua be
a meta-port that depends on all kyua packages.

Because there is no version number to represent Kyua as a whole (that's
a problem I have to resolve upstream somehow), bump the epoch of
devel/kyua and use today's date as the version.

No entry should be necessary in MOVED because upgrading to the new
meta-port should do the right thing.

Approved by:	bdrewery


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344497 35697150-7ecd-e111-bb59-0022644237b5