devel/sgl: Upon inspection, port is inferior - deprecate
After this Ada port was returned to the heap, I took a look at it.  The
primary issue is that it's just copying source files.  It should be
building a library in the standard location as well.  That's easily
fixable except there's no point.

This software was just a graduate thesis from Spring 1996 that was
never developed beyond that.  It wasn't complete at the time.  As far
as I know, there isn't even a demand for having an STL port in Ada95.
Later versions of Ada have extensive libraries that basically do what
SGL was attempting to do.  Original PR didn't explain why the initiator
thought SGL was necessary or useful 6 years later.  Just kill it,
expiration in a month.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344506 35697150-7ecd-e111-bb59-0022644237b5