devel/ahven: Modify to add symbolic tracing.
The author, Tero Koskine, intends for ahven to be compiler agnostic,
and therefore can't use GNAT-specific functionality like
GNAT.Traceback.Symbolic.  However, there is only one brand of Ada
compiler in ports, GNAT, so we can take advantage of it.

This change adds a default option to alter the exception handling to
trace through the symbols to display files and line numbers of each
frame on the stack.  Declining the option builds Ahven without the
modification and allows Ahven to be used on non-GNAT compilers (assuming
you can find one on FreeBSD.)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@344938 35697150-7ecd-e111-bb59-0022644237b5