sysutils/dd_rescue: unbreak everywhere
The internal makefile calls autoconf and autoheader, but neither
were specified as build dependencies before now.

Note that this port will not build on FreeBSD 8 because __WORDSIZE is not
defined in _stdint.h yet.  DragonFly got around it with setting these
in Makefile.DragonFly:

  x86_64:
    MAKE_ARGS+= CFLAGS_OPT="-D__WORDSIZE=64"
    CFLAGS+=    -D__WORDSIZE=64

  i386:
    MAKE_ARGS+= CFLAGS_OPT="-D__WORDSIZE=32"
    CFLAGS+=    -D__WORDSIZE=32

A similar fix could be done for FreeBSD 8, but it needs to cover more
architectures.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@345158 35697150-7ecd-e111-bb59-0022644237b5