net/nakenchat: upgrade from 2.12 to 2.22 and unbreak
Nakenchat has been broken everywhere for some kind of installation issue.
Adding stage support fixes that issue, but let's update to the latest
stable version of 2.22 while here.  A license was also added.  Per a
pending PR, nakenpasswd was removed.  It's now a bash script rather than
a compiled program, but I guess the maintainer thought putting bash as
a run dependency wasn't worth it.  Most of the PR was used, but it
requested an upgrade to a beta version 3.00b.  I'm holding off on that
upgrade pending feedback.

PR:	ports/185440


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@345486 35697150-7ecd-e111-bb59-0022644237b5