Add new port net/anet (Ada IPv4 and IPv6 sockets binding library)
The ANet library was created on Linux and the unfortunate result is that
it is highly Linux-specific.  Luckily it has an implementation testsuite,
so patches to make it work on BSD can be reasonably tested.  The current
status is annoted in pkg-message:

=========================================================================
Beware of the IPv6 multicast functions.  Sending does work, but the
default interface effectively is invalid on *BSD.  A specific interface
needs to be provided rather than leaving interface blank (zero).

Multicast receiving may not currently work.  The test for IPv6 multicast
fails.  The test chunk is sent (verified with separate monitoring tool)
but never detected.  Hopefully the cause will be understood and fixed
soon.

AF_NETLINK and AF_PACKET protocols are not supported by *BSD, so the
associated tests have been removed.  Every test other than IPv6 Multicast
passes.  You may want to replace "em0" with this system's interface in
the test suite is to be run (see files/patch-tests_socket__tests.adb).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@346016 35697150-7ecd-e111-bb59-0022644237b5