x11-wm/stumpwm: Failed to restore broken port, mark for deletion in 2 months
This port supposedly builds with two different lisp compilers.  The
default one caused a BROKEN definition, so the fix seemed simple: remove
the default and let the other compiler build it.

It didn't work.  The other compiler requires a non-default option (new clx)
meaning it wouldn't build in poudriere, and moreover the test for it was
broken.  I fixed the test, but the makefile looks for lisp.run in the
wrong place.

I believe clisp could build this unmaintained port with sufficient patches,
but they won't come from me.  I started to clean the port up a bit, and
I left those improvements in place.  Somebody else wil need to need to
restore SBCL support and/or CLISP support within 2 months, otherwise this
port is headed to the garbage collector.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@346083 35697150-7ecd-e111-bb59-0022644237b5