devel/capstone: Add DragonFly support
The makefile patch was strange in that it defined LIBDATADIR once
unconditionally and then again conditionally.  This could have been
done via MAKE_ARGS and no patch at all.  However, I changed the
patch to a single definition that is true for "ports" regardless
of the platform used (pkgconfig files always in libdata in ports).
This allows port to build on DragonFly and no-op for FreeBSD.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@346753 35697150-7ecd-e111-bb59-0022644237b5