lang/gcc-aux: Fix skip-bootstrap functionality
I don't know when the skip-bootstrap logic stopped working, but I'm
reasonably sure it did once.  Apparently PREFIX isn't defined at the
time of the "if exists" evaluation.  Using LOCALBASE instead will allow
the compiler to build using the compiler on the system rather than
the older bootstrap compiler.  The bug was discovered while trying to
move gcc-aux to a gcc49 base.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@346972 35697150-7ecd-e111-bb59-0022644237b5