MFH: r347413
Switch to using Google's tarball for both the browser and the test data.
This removes one chain in the link of trust since the tarball is no longer
modified and rerolled locally.

This means the Courgette code is now included, but it is only compiled and
used on Windows platforms to distribute binary updates.

Legal documentation:
http://law.justia.com/cases/federal/district-courts/massachusetts/madce/1:2009cv11813/125212/47/

Discussed with: portmgr/core (tabthorpe), FreeBSD Foundation (gnn)

Approved by:	portmgr (tabthorpe)


git-svn-id: svn+ssh://svn.freebsd.org/ports/branches/2014Q1@347424 35697150-7ecd-e111-bb59-0022644237b5