Add USES=gmake to fix installation on systems that use fmake by default.
Some makefiles in this port list "configure" as a dependency of some files.
Because "configure" gets modified by USES=libtool, this means make tries to
regenerate them.  The commands for these files detect that nothing has
changed however and they don't actually regenerate the files.  Both bmake
and gmake leave it at that, but fmake reruns all targets that depend on
these files anyway.  This fails because it requires tools that aren't
available.

Reported by:	antoine


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@347528 35697150-7ecd-e111-bb59-0022644237b5