MozJPEG is a fork of libjpeg-turbo with 'jpgcrush' functionality built in.
The goal is to provide a production-quality JPEG encoder that improves
compression while maintaining compatibility with the vast majority of deployed
decoders.

The 'jpgcrush' feature finds the progressive coding configuration which uses the
fewest bits. This most frequently reduces file size by 2-10%, but those are not
hard limits. Significantly greater reductions have been observed.

Library configuration defaults are the same as for libjpeg-turbo, in order to
make transitions as painless as possible. There are new configuration options
for new features, but they are not enabled by default.

The 'cjpeg' program defaults are not the same as for the equivalent program in
libjpeg-turbo. The 'cjpeg' defaults for mozjpeg are set to aggressively optimize
for smaller file sizes.

WWW: https://github.com/mozilla/mozjpeg/

PR:		ports/187398
Submitted by:	Horia Racoviceanu <horia@racoviceanu.com>


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@348122 35697150-7ecd-e111-bb59-0022644237b5