- Add imageworsener 1.2.0
ImageWorsener is a cross-platform command-line utility and library for image
scaling and other image processing. It has full support for PNG, JPEG, BMP
formats, experimental support for WebP, read-only support for GIF, and limited
support for some other image formats. It's not as fast or memory-efficient as
some utilities, but it's very accurate.

WWW: http://entropymine.com/imageworsener/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@348372 35697150-7ecd-e111-bb59-0022644237b5