Add new Ada port: textproc/xml_ez_out
XML EZ_Out is a small set of packages intended to aid the creation of
XML-formatted output from within Ada programs.  It basically wraps the
tags and data provided to it with XML syntax and writes them to a
user-supplied medium.

This medium can be any sort of writable entity, such as a file, a
memory buffer, or even a communications link, such as a socket.  The
only functionality required of the medium is that it supply a
meaningful "Put" (for writing a string) and "New_Line" procedure.

WWW: http://www.mckae.com/xmlEz.html


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@348387 35697150-7ecd-e111-bb59-0022644237b5