devel/ocaml-opam: Update 1.1.0 -> 1.1.1 and unbreak
Updating to the latest point version did not fix the problems.  The
first is that now the built tool ocp-build requires TERM to be set in
the environment, otherwise it returns non-zero result with an error
message (hidden by /dev/null in first instance) and breaks the build.

Later the build failed the FS because it was saving preferences at $HOME.
Adding -no-user-prefs switch to the command seems to suppress that.

Finally, the GNU makefile is not jobs safe, so indicate that.  The GNU
makefile does not participate much in the overall build.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@349354 35697150-7ecd-e111-bb59-0022644237b5