lang/gcc-aux: Add binutils requirement for FreeBSD
This patch was previously developed but lost somewhere along the way.
The linker and assembly must be explicitly defined otherwise it is
defined as "ld" and "as" and as such is affected by the value of PATH.
We don't want this, and it actually breaks the gnatdroid cross compilers.

For FreeBSD, use the binutils port like the gcc ports do.  DragonFly
has binutils 2.24 in base so let's keep using the base binutils on that
platform.

Also the Makefile.version file now defines the PORTREVISION for all
the ports that use it to make future updating easier.

Binutils patch omission partly detected by jenkins failure (thx swills)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@350787 35697150-7ecd-e111-bb59-0022644237b5