lang/gnatdroid-armv7 (&gcc-aux): Fix socket support and gnat.dg testsuite
The socket support on Gnatdroid was broken, and it had to be fixed at
lang/gcc-aux.

Until now, the dejagnu testsuite did not work on remote targets.  To fix
this, support files had to be installed at ${LOCALBASE}/share/dejagnu.
Dejagnu was not made a run dependency, conditional or otherwise.  The
test targets for gnatdroid are really maintainer targets.

Running the gnat.dg testsuite confirms that everything works in the
cross-compiler except stack checking (there's no support in gcc for this
just yet).


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@351148 35697150-7ecd-e111-bb59-0022644237b5