databases/sqlbuddy: Unbreak and stage
The maintainer actually provided a PR to address the breakage long
before it was marked broken.  itetcu asked him a valid question: Why
is sqlbuddy defaulting to sqlite when it is primarily a mysql tool?
The question was never answered and the PR got thrown back on the heap.

The SQLite component is what is missing.  Let's remove sqlite as an
option to restore the build and stage while here.  It's bumped because
of the change in default options.

PR:		ports/181708
submitted by:	maintainer (Chenguang Li)
modified by:	marino
MFH:		2014Q2


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@351703 35697150-7ecd-e111-bb59-0022644237b5