Add new Ada port: textproc/templates_parser
This was part of the Ada Web Server, but it is maintained separated and
even needs to be recursively cloned into AWS git repository.  Now the
AWS template engine is a separate library that AWS will use in the
upcoming update.  The package description is below.

     ================================================================

This is the template engine for the Ada Web Server.  It is modular and
therefore can be split out of AWS and used on its own.

As it was designed for generating web pages, it's function is to parse
a page template and replace tokens with specified values.  This template
engine is amazingly fast due to its concurrent cached compiled templates
support.

WWW: http://docs.adacore.com/aws-docs/templates_parser


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@354249 35697150-7ecd-e111-bb59-0022644237b5