Change ports that use the ATF libraries to depend on devel/atf.
As part of this, make it possible to install devel/atf independently of
the presence of ATF in the base system.  This is so that we can turn the
ATF libraries in base into private libraries.

Note that this should have no effect on prebuilt binary packages because
these changes only apply when the TEST option is selected, and this option
is not yet the default.

Approved by:	eadler (ports)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355077 35697150-7ecd-e111-bb59-0022644237b5