Enable the TEST option by default in atf-based ports.
Get ports that use the atf libraries to install their tests by default.
This is to showcase the potential of getting the tests installed on a
live system and will also simplify the process of running them periodically
out of /usr/local/tests/.

Packages confirmed to build with this option enabled.

Approved by:	bapt (ports)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@355276 35697150-7ecd-e111-bb59-0022644237b5