graphics/darktable: Update to 1.4.2
The following changes were made to options:

    o  Remove GCC option (see below)
    o  Remove KWALLET option (unused)
    o  Rename FACEBOOK option to FB_PICASA, because both export features
       depend on json-glib.
    o  Add COLORD option (color management)
    o  Add GRAPHMAGICK option (support GraphicsMagick's image formats)
    o  Add LUA option (embed Lua to add scripting)
    o  Add OPENEXR option (support HDR image format)
    o  Add SQUISH option (use libsquish to compress thumbnails)
    o  Add WEBP option (support WebP image format)

The patch enables more options by default, because I think it's more
sensible for a photographer using binary packages:

    o  COLORD
    o  FB_PICASA
    o  FLICKR
    o  GEO
    o  GPHOTO (already on)
    o  LUA
    o  NLS
    o  OPENEXR
    o  OPENJPEG
    o  RAWSPEED (already on)
    o  WEBP

The GCC option was removed because the usage of GCC along with libc++
in FreeBSD 10 and -CURRENT is unstable (this isn't a general truth,
just valid in the case of darktable). In FreeBSD up-to 9, the usage of
lang/gcc is mandatory anyway, because GCC 4.2.1 support is best effort
only (upstream).

Ninja is now used to build darktable, instead of make. This is not mandatory
of course, but brings a 33% build time improvement (at least for me).

pkg-plist is generated with "make makeplist" this time. That's why the
diff seems large. In fact most of the content remains the same, just with
a different order.

All patches were committed upstream and are removed, except one which
comes from upstream but isn't part of release 1.4.2.

The port is now "stagified".

Finally, I take maintainership of it.

PR:		ports/186979
Phabricator:	D107
Reviewed by:	danfe@ (previous version), antoine@
Approved by:	antoine@


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@356162 35697150-7ecd-e111-bb59-0022644237b5