MFH: r357000
Set PYTHON_PY3K_PLIST_HACK to adjust the plist for Python >= 3.2.0.

Since this port does not use distutils/setuptools and we manually generate
the .pyc/.pyo files, we need to set PYTHON_PY3K_PLIST_HACK so that the cache
files end up in the right directory if Python >= 3.2.0 is used due to
PEP-3147.

PR:		188150

Approved by:	portmgr (miwi)


git-svn-id: svn+ssh://svn.freebsd.org/ports/branches/2014Q2@357002 35697150-7ecd-e111-bb59-0022644237b5