- Add new port sysutils/pstacku using libunwind and written by kib. - Add CONFLICTS with sysutils/pstack.
"pstack" is a workalike for the Solaris program of the same name.
Running pstack on a process produces a stack trace of
each thread in that process.

It's useful for finding out what wedged processes are up to, getting
profiles of applications, and just satisfying one's curiosity.

WWW: https://www.kib.kiev.ua/git/gitweb.cgi?p=pstacku.git;a=tree


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@358295 35697150-7ecd-e111-bb59-0022644237b5