graphics/gdchart: Replace wrong PROFILE option
The profile library gets built if NO_PROFILE is not set.  All the
PROFILE option was doing was manually changing the pkg-plist which
would only work if the PROFILE option was manually synchronized with
the make.conf option (they are by default, so it worked accidently).

Change PLIST_SUB to switch based on NO_PROFILE as is done with other
ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@360798 35697150-7ecd-e111-bb59-0022644237b5