irc/nefarious: stage, take 2.  Add DOCS options
The vendor makefile creates html documents unconditionally with two
flavors: operators and users.  There is a lot of duplication between the
two so it symlinked the duplicates.  Unfortunately the symlinks point
back the stage directory.  Also, at least one symlink was invalid in all
cases as it was not pointing to any file.

To fix:
 1. override install target
 2. Add DOCS option to make non-man page documentation optional
 3. Copy files instead of using symlinks (they are very small)
 4. Fix vendor SYMLINK list, there were two errors in it
 5. Remove subversion tree during post-extract target

Note that the permissions of the installed configuration files are
different.  The vendor files tried to install them with modes 600 and 700
but now they are the default for INSTALL_DATA macro.

Reported by:	swills (jenkins)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@360864 35697150-7ecd-e111-bb59-0022644237b5