games/atomix: Fix deinstallation issue involving scores file
The scheme used to leave a non-zero length scores file alone during
deinstallation no longer works under stage.  The scheme used a
post-install script to add the scores directory and initial file, but
then tried to remove it using the pkg-plist.

Apparently these need to be matched, which means either the post-install
script is matched with a post-deinstall script, or the creation of the
scores file be done in the pkg-list with @exec.  As the logic is slightly
complex, it was easier to solve with a pkg-deinstall script.

There were two unnecessary @dirrmtry lines as well which have been
removed.  Bump portrevision for multiple reasons.

Reported by:	swills (jenkins)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@361273 35697150-7ecd-e111-bb59-0022644237b5