Remove editors/nxml because it needs recently removed emacs 21 or 22
The nxml editor sets IGNORE for emacs23 and emacs24 since its
functionality is incorporated into those versions of emacs.  It was
meant to be used with emacs21 or emacs22, but since those ports were
recently removed, IGNORE is set in all cases.

This port is unstaged, so it would have been reset and deprecated in any
case soon.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@361687 35697150-7ecd-e111-bb59-0022644237b5