New port: devel/p5-IO-Prompter:
IO::Prompter exports a single subroutine, prompt, that prints a prompt (but
only if the program's selected input and output streams are connected to a
terminal), then reads some input, then chomps it, and finally returns an
object representing that text.

IO::Prompter is a replacement for IO::Prompt, which is no longer being maintained.

WWW: http://search.cpan.org/dist/IO-Prompter/


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@362100 35697150-7ecd-e111-bb59-0022644237b5