Add new port www/vimb
PR:		191804
Submitted by:	mike.d.ft402 (gmail.com)

Vimb is a web browser that behaves like Vimprobable, but with some
paradigms from dwb and hopefully a cleaner code base. The goal of Vimb is
to build a completely keyboard-driven, efficient and pleasurable browsing
experience with low memory and cpu usage.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@364484 35697150-7ecd-e111-bb59-0022644237b5