Reset maintainership on www/dwoo
This port was not staged, but it was spared a reset because somebody (not
the maintainer) submitted a PR to stage it.  Unfortunately the PR has
numerous problems which have not been addressed in the last couple of
weeks.  Reset the maintainer now, and let the port be pruned at the end
of the month if no one steps up to save it.

PR:		190041
Submitted by:	Joe (thralling.com)
Approved by:	maintainer (Joe Horn)
Rejected by:	marino (technical reasons)
Approved by:	portmgr (implicit, NOT_STAGED)


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@364543 35697150-7ecd-e111-bb59-0022644237b5