textproc/ibus: Removing unnecessary @dirrmtry from pkg-plist
This port failed to build with poudriere bulk -t because the directories
it tried to remove were considered part of MTREE (but its seems they only
belonged to dependencies to me).

Remove the entries from pkg-plist so it builds cleanly, but don't bother
to revbump because there's no practical change to the package.
Just Fix It.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@365351 35697150-7ecd-e111-bb59-0022644237b5